requirements.txt
Flask==3.0.0
numpy==1.26.2
numpy-stl==3.1.1
anthropic==0.39.0
werkzeug==3.0.1
